// src/api/generated/public/schema.ts
export interface components {
    schemas: {
// ...
        PublicContent: {
            aeo: components["schemas"]["PublicAEO"];
            alternates: components["schemas"]["PublicAlternate"][];
            blocks: components["schemas"]["PublicBlock"][];
            /** Format: uuid */
            content_id: string;
            content_key: string;
            /** @enum {string} */
            content_status: "published";
            content_version: number;
            excerpt: string;
            geo: components["schemas"]["PublicGEO"];
            /** @enum {string} */
            kind: "page" | "post";
            /** Format: uuid */
            locale_id: string;
            path: string;
            /** Format: date-time */
            published_at?: string;
            seo: components["schemas"]["PublicSEO"];
            slug: string;
            title: string;
            /** Format: uuid */
            translation_id: string;
            /** @enum {string} */
            translation_status: "published";
            translation_version: number;
            /** Format: date-time */
            updated_at: string;
        };
// ...
    };
// ...
}